chore: more progress
