+++ src/gfx/draw/drawBezier.ts
@@ -2,6 +2,10 @@ import { evaluateBezier } from "../../math/math";
 import { type Vec2 } from "../../math/Vec2";
 import { drawCurve, type DrawCurveOpt } from "./drawCurve";
 
+/**
+ * @group Draw
+ * @subgroup Types
+ */
 export type DrawBezierOpt = DrawCurveOpt & {
     /**
      * The first point.
