+++ src/gfx/draw/drawUVQuad.ts
@@ -1,9 +1,10 @@
 import { DEF_ANCHOR, UV_PAD } from "../../constants/general";
-import { Color, rgb } from "../../math/color";
+import { Color } from "../../math/color";
 import { Quad } from "../../math/math";
 import { Vec2 } from "../../math/Vec2";
-import { BlendMode, type DrawUVQuadOpt } from "../../types";
+import { type Anchor, BlendMode, type RenderProps } from "../../types";
 import { anchorPt } from "../anchor";
+import type { Texture } from "../gfx";
 import {
     multRotate,
     multScaleV,
@@ -13,6 +14,43 @@ import {
     pushTransform,
 } from "../stack";
 import { drawRaw } from "./drawRaw";
+
+/**
+ * How the UV Quad should look like.
+ *
+ * @group Draw
+ * @subgroup Types
+ */
+export type DrawUVQuadOpt = RenderProps & {
+    /**
+     * Width of the UV quad.
+     */
+    width: number;
+    /**
+     * Height of the UV quad.
+     */
+    height: number;
+    /**
+     * If flip the texture horizontally.
+     */
+    flipX?: boolean;
+    /**
+     * If flip the texture vertically.
+     */
+    flipY?: boolean;
+    /**
+     * The texture to sample for this quad.
+     */
+    tex?: Texture;
+    /**
+     * The texture sampling area.
+     */
+    quad?: Quad;
+    /**
+     * The anchor point, or the pivot point. Default to "topleft".
+     */
+    anchor?: Anchor | Vec2;
+};
 
 export function drawUVQuad(opt: DrawUVQuadOpt) {
     if (opt.width === undefined || opt.height === undefined) {
